refactor mesh components from both fem and bem into top-level mesh module
