+++ src/mesh/elements/mod.rs
@@ -1,0 +1,7 @@
+//! Reference-element definitions grouped by element family.
+
+pub mod quad2d;
+pub mod tri;
+
+pub use quad2d::{quad4, quad9};
+pub use tri::tri3;
