+++ src/mesh/elements/quad2d/mod.rs
@@ -1,0 +1,8 @@
+//! Quadrilateral reference elements and shared quad-family helpers.
+
+pub mod mapping;
+pub mod quad4;
+pub mod quad9;
+pub mod quadrature;
+
+pub use quadrature::QuadratureRule;
