+++ src/mesh/elements/tri/mod.rs
@@ -1,0 +1,5 @@
+//! Triangle reference elements and shared triangle-family helpers.
+
+pub mod mapping;
+pub mod quadrature;
+pub mod tri3;
