+++ src/physics/solenoid_stress/assembly.rs
@@ -4,23 +4,22 @@
 //! `K_e = integral(B^T D B 2*pi*r dA)` and the consistent load vectors are assembled from the
 //! same weak form.
 
+use crate::mesh::elements::quad2d::{quad4, quad9};
+use crate::mesh::{MeshView, QuadratureRule};
 use crate::physics::solenoid_stress::axisym::{
     accumulate_b_transpose_vector, accumulate_stiffness, build_b_matrix, constitutive_times_strain,
 };
 use crate::physics::solenoid_stress::geometry::{
-    VolumeSample, volume_samples_quad4, volume_samples_quad9,
+    VolumeSample, validate_axisymmetric_nodes, volume_samples_quad4, volume_samples_quad9,
 };
 use crate::physics::solenoid_stress::loads::{
     accumulate_body_force, pressure_element_load_quad4, pressure_element_load_quad9,
     traction_element_load_quad4, traction_element_load_quad9,
 };
-use crate::physics::solenoid_stress::mesh::{
-    AssemblyResult, MeshView, PressureLoad, ThermalMaterial, TractionLoad,
+use crate::physics::solenoid_stress::types::{
+    AssemblyResult, DOF_PER_NODE, PressureLoad, Real, ThermalMaterial, TractionLoad,
+    dof_per_element, two_pi,
 };
-use crate::physics::solenoid_stress::quad4;
-use crate::physics::solenoid_stress::quad9;
-use crate::physics::solenoid_stress::quadrature::QuadratureRule;
-use crate::physics::solenoid_stress::types::{Real, two_pi};
 
 fn assemble_axisymmetric_impl<
     F: Real,
@@ -53,8 +52,8 @@ fn assemble_axisymmetric_impl<
         QuadratureRule,
     ) -> Result<[F; DOF_PER_ELEMENT], String>,
 ) -> Result<AssemblyResult<F>, String> {
-    debug_assert_eq!(DOF_PER_ELEMENT, 2 * NODES_PER_ELEMENT);
-    mesh.validate_nodes()?;
+    debug_assert_eq!(DOF_PER_ELEMENT, DOF_PER_NODE * NODES_PER_ELEMENT);
+    validate_axisymmetric_nodes(mesh)?;
     mesh.validate_connectivity()?;
     if material_ids.len() != mesh.num_elements() {
         return Err(format!(
@@ -216,7 +215,11 @@ pub fn assemble_axisymmetric_quad4<F: Real>(
     nodal_temperature: Option<&[F]>,
     quadrature: QuadratureRule,
 ) -> Result<AssemblyResult<F>, String> {
-    assemble_axisymmetric_impl(
+    assemble_axisymmetric_impl::<
+        F,
+        { quad4::NODES_PER_ELEMENT },
+        { dof_per_element(quad4::NODES_PER_ELEMENT) },
+    >(
         mesh,
         material_ids,
         material_table,
@@ -244,7 +247,11 @@ pub fn assemble_axisymmetric_quad9<F: Real>(
     nodal_temperature: Option<&[F]>,
     quadrature: QuadratureRule,
 ) -> Result<AssemblyResult<F>, String> {
-    assemble_axisymmetric_impl(
+    assemble_axisymmetric_impl::<
+        F,
+        { quad9::NODES_PER_ELEMENT },
+        { dof_per_element(quad9::NODES_PER_ELEMENT) },
+    >(
         mesh,
         material_ids,
         material_table,
@@ -263,8 +270,7 @@ pub fn assemble_axisymmetric_quad9<F: Real>(
 #[cfg(test)]
 mod tests {
     use super::assemble_axisymmetric_quad4;
-    use crate::physics::solenoid_stress::mesh::MeshView;
-    use crate::physics::solenoid_stress::quadrature::QuadratureRule;
+    use crate::mesh::{MeshView, QuadratureRule};
 
     fn isotropic_material(e: f64, nu: f64) -> [[f64; 4]; 4] {
         let lam = e * nu / ((1.0 + nu) * (1.0 - 2.0 * nu));
