+++ src/physics/solenoid_stress/geometry.rs
@@ -1,13 +1,15 @@
-//! Geometry and quadrature helpers for axisymmetric quadrilateral elements.
+//! Axisymmetric wrappers around generic 2D quadrilateral geometry sampling.
 //!
-//! These routines sit between the purely reference-element formulas in the element modules and
-//! the assembly kernels. Their job is to evaluate everything at physical quadrature points:
-//! position `(r, z)`, Jacobian-derived weights, and shape-function gradients in physical space.
+//! The reusable Jacobian, mapping, and quadrature sampling logic lives in [`crate::mesh`]. This
+//! module adds the axisymmetric-specific validation and the `2*pi*r`-weighted element summaries
+//! needed by the structural solver.
 
-use crate::physics::solenoid_stress::mesh::MeshView;
-use crate::physics::solenoid_stress::quadrature::{QuadratureRule, gauss_face, gauss_volume};
+use crate::mesh::elements::quad2d::{quad4, quad9};
+use crate::mesh::sampling;
+use crate::mesh::{MeshView, QuadratureRule};
 use crate::physics::solenoid_stress::types::{Real, two_pi};
-use crate::physics::solenoid_stress::{quad4, quad9};
+
+pub use crate::mesh::{FaceSample, VolumeSample};
 
 #[derive(Debug, Clone)]
 pub struct ElementMeasures<F: Real> {
@@ -29,176 +31,60 @@ pub struct ElementQuadrature<F: Real> {
     pub nq_per_element: usize,
 }
 
-#[derive(Debug, Clone, Copy)]
-pub struct VolumeSample<F: Real, const NODES_PER_ELEMENT: usize> {
-    /// Shape-function values at the quadrature point.
-    pub n: [F; NODES_PER_ELEMENT],
-    /// Physical gradients `[dN_i/dr, dN_i/dz]`.
-    pub grad_phys: [[F; 2]; NODES_PER_ELEMENT],
-    /// Jacobian determinant `det(J)` of the reference-to-physical map.
-    pub det_j: F,
-    /// Physical quadrature point `(r, z)`.
-    pub point: [F; 2],
-    /// Reference-space quadrature weight `w`.
-    pub weight: F,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct FaceSample<F: Real, const NODES_PER_ELEMENT: usize> {
-    /// Shape-function values on the element face.
-    pub n: [F; NODES_PER_ELEMENT],
-    /// Physical tangent vector corresponding to the reference edge direction.
-    pub tangent: [F; 2],
-    /// Physical quadrature point `(r, z)` on the face.
-    pub point: [F; 2],
-    /// Reference-edge quadrature weight.
-    pub weight: F,
-}
-
-fn map_point<F: Real, const NODES_PER_ELEMENT: usize>(
-    coords: &[[F; 2]; NODES_PER_ELEMENT],
-    n: &[F; NODES_PER_ELEMENT],
-) -> [F; 2] {
-    let mut point = [F::zero(); 2];
-    for i in 0..NODES_PER_ELEMENT {
-        point[0] = point[0] + n[i] * coords[i][0];
-        point[1] = point[1] + n[i] * coords[i][1];
+pub(crate) fn validate_axisymmetric_nodes<F: Real, const NODES_PER_ELEMENT: usize>(
+    mesh: MeshView<'_, F, NODES_PER_ELEMENT>,
+) -> Result<(), String> {
+    for (index, node) in mesh.nodes_rz.iter().enumerate() {
+        if node[0] < F::zero() {
+            return Err(format!(
+                "node {index} has negative radius {:?}; axisymmetric radius must be nonnegative",
+                node[0]
+            ));
+        }
     }
-    point
+    Ok(())
 }
 
-fn jacobian<F: Real, const NODES_PER_ELEMENT: usize>(
-    coords: &[[F; 2]; NODES_PER_ELEMENT],
-    grad: &[[F; 2]; NODES_PER_ELEMENT],
-) -> [[F; 2]; 2] {
-    let mut jac = [[F::zero(); 2]; 2];
-    for i in 0..NODES_PER_ELEMENT {
-        jac[0][0] = jac[0][0] + coords[i][0] * grad[i][0];
-        jac[0][1] = jac[0][1] + coords[i][0] * grad[i][1];
-        jac[1][0] = jac[1][0] + coords[i][1] * grad[i][0];
-        jac[1][1] = jac[1][1] + coords[i][1] * grad[i][1];
-    }
-    jac
-}
-
-fn det_j<F: Real>(jac: &[[F; 2]; 2]) -> F {
-    jac[0][0] * jac[1][1] - jac[0][1] * jac[1][0]
-}
-
-fn inv_j<F: Real>(jac: &[[F; 2]; 2]) -> Result<[[F; 2]; 2], String> {
-    let det = det_j(jac);
-    if det <= F::zero() {
-        return Err(format!(
-            "encountered non-positive element Jacobian determinant {det:?}"
-        ));
-    }
-    let inv_det = F::one() / det;
-    Ok([
-        [jac[1][1] * inv_det, -jac[0][1] * inv_det],
-        [-jac[1][0] * inv_det, jac[0][0] * inv_det],
-    ])
-}
-
-fn grad_phys<F: Real, const NODES_PER_ELEMENT: usize>(
-    grad_reference: &[[F; 2]; NODES_PER_ELEMENT],
-    inv_jac: &[[F; 2]; 2],
-) -> [[F; 2]; NODES_PER_ELEMENT] {
-    let mut out = [[F::zero(); 2]; NODES_PER_ELEMENT];
-    for i in 0..NODES_PER_ELEMENT {
-        let dxi = grad_reference[i][0];
-        let deta = grad_reference[i][1];
-        out[i][0] = inv_jac[0][0] * dxi + inv_jac[1][0] * deta;
-        out[i][1] = inv_jac[0][1] * dxi + inv_jac[1][1] * deta;
-    }
-    out
-}
-
-fn volume_samples_generic<F: Real, const NODES_PER_ELEMENT: usize>(
-    coords: &[[F; 2]; NODES_PER_ELEMENT],
-    quadrature: QuadratureRule,
-    shape_fn: fn(F, F) -> [F; NODES_PER_ELEMENT],
-    grad_ref_fn: fn(F, F) -> [[F; 2]; NODES_PER_ELEMENT],
+fn validate_axisymmetric_volume_samples<F: Real, const NODES_PER_ELEMENT: usize>(
+    samples: Vec<VolumeSample<F, NODES_PER_ELEMENT>>,
 ) -> Result<Vec<VolumeSample<F, NODES_PER_ELEMENT>>, String> {
-    let samples = gauss_volume::<F>(quadrature);
-    let mut out = Vec::with_capacity(samples.len());
-    for ([xi, eta], weight) in samples {
-        let n = shape_fn(xi, eta);
-        let grad_reference = grad_ref_fn(xi, eta);
-        let jac = jacobian(coords, &grad_reference);
-        let inv = inv_j(&jac)?;
-        let det = det_j(&jac);
-        let point = map_point(coords, &n);
-        if point[0] < F::zero() {
+    for sample in &samples {
+        if sample.point[0] < F::zero() {
             return Err(format!(
                 "quadrature point has negative radius {:?}; axisymmetric radius must be nonnegative",
-                point[0]
+                sample.point[0]
             ));
         }
-        out.push(VolumeSample {
-            n,
-            grad_phys: grad_phys(&grad_reference, &inv),
-            det_j: det,
-            point,
-            weight,
-        });
     }
-    Ok(out)
+    Ok(samples)
 }
 
-fn face_samples_generic<F: Real, const NODES_PER_ELEMENT: usize>(
-    coords: &[[F; 2]; NODES_PER_ELEMENT],
-    local_face: u8,
-    quadrature: QuadratureRule,
-    shape_fn: fn(F, F) -> [F; NODES_PER_ELEMENT],
-    grad_ref_fn: fn(F, F) -> [[F; 2]; NODES_PER_ELEMENT],
-    face_ref_fn: fn(u8, F) -> Result<(F, F, [F; 2]), String>,
+fn validate_axisymmetric_face_samples<F: Real, const NODES_PER_ELEMENT: usize>(
+    samples: Vec<FaceSample<F, NODES_PER_ELEMENT>>,
 ) -> Result<Vec<FaceSample<F, NODES_PER_ELEMENT>>, String> {
-    let samples = gauss_face::<F>(quadrature);
-    let mut out = Vec::with_capacity(samples.len());
-    for (s, weight) in samples {
-        let (xi, eta, ds_reference) = face_ref_fn(local_face, s)?;
-        let n = shape_fn(xi, eta);
-        let grad_reference = grad_ref_fn(xi, eta);
-        let jac = jacobian(coords, &grad_reference);
-        let point = map_point(coords, &n);
-        if point[0] < F::zero() {
+    for sample in &samples {
+        if sample.point[0] < F::zero() {
             return Err(format!(
                 "face quadrature point has negative radius {:?}; axisymmetric radius must be nonnegative",
-                point[0]
+                sample.point[0]
             ));
         }
-        let tangent = [
-            jac[0][0] * ds_reference[0] + jac[0][1] * ds_reference[1],
-            jac[1][0] * ds_reference[0] + jac[1][1] * ds_reference[1],
-        ];
-        let tangent_norm_sq = tangent[0] * tangent[0] + tangent[1] * tangent[1];
-        if tangent_norm_sq <= F::zero() {
-            return Err(format!(
-                "degenerate face tangent on local face {local_face}; tangent squared norm is {tangent_norm_sq:?}"
-            ));
-        }
-        out.push(FaceSample {
-            n,
-            tangent,
-            point,
-            weight,
-        });
     }
-    Ok(out)
+    Ok(samples)
 }
 
 pub fn volume_samples_quad4<F: Real>(
     coords: &[[F; 2]; quad4::NODES_PER_ELEMENT],
     quadrature: QuadratureRule,
 ) -> Result<Vec<VolumeSample<F, { quad4::NODES_PER_ELEMENT }>>, String> {
-    volume_samples_generic(coords, quadrature, quad4::shape::<F>, quad4::grad_ref::<F>)
+    validate_axisymmetric_volume_samples(sampling::volume_samples_quad4(coords, quadrature)?)
 }
 
 pub fn volume_samples_quad9<F: Real>(
     coords: &[[F; 2]; quad9::NODES_PER_ELEMENT],
     quadrature: QuadratureRule,
 ) -> Result<Vec<VolumeSample<F, { quad9::NODES_PER_ELEMENT }>>, String> {
-    volume_samples_generic(coords, quadrature, quad9::shape::<F>, quad9::grad_ref::<F>)
+    validate_axisymmetric_volume_samples(sampling::volume_samples_quad9(coords, quadrature)?)
 }
 
 pub fn face_samples_quad4<F: Real>(
@@ -206,14 +92,9 @@ pub fn face_samples_quad4<F: Real>(
     local_face: u8,
     quadrature: QuadratureRule,
 ) -> Result<Vec<FaceSample<F, { quad4::NODES_PER_ELEMENT }>>, String> {
-    face_samples_generic(
-        coords,
-        local_face,
-        quadrature,
-        quad4::shape::<F>,
-        quad4::grad_ref::<F>,
-        quad4::face_reference::<F>,
-    )
+    validate_axisymmetric_face_samples(sampling::face_samples_quad4(
+        coords, local_face, quadrature,
+    )?)
 }
 
 pub fn face_samples_quad9<F: Real>(
@@ -221,14 +102,9 @@ pub fn face_samples_quad9<F: Real>(
     local_face: u8,
     quadrature: QuadratureRule,
 ) -> Result<Vec<FaceSample<F, { quad9::NODES_PER_ELEMENT }>>, String> {
-    face_samples_generic(
-        coords,
-        local_face,
-        quadrature,
-        quad9::shape::<F>,
-        quad9::grad_ref::<F>,
-        quad9::face_reference::<F>,
-    )
+    validate_axisymmetric_face_samples(sampling::face_samples_quad9(
+        coords, local_face, quadrature,
+    )?)
 }
 
 fn element_measures_generic<F: Real, const NODES_PER_ELEMENT: usize>(
@@ -239,7 +115,7 @@ fn element_measures_generic<F: Real, const NODES_PER_ELEMENT: usize>(
         QuadratureRule,
     ) -> Result<Vec<VolumeSample<F, NODES_PER_ELEMENT>>, String>,
 ) -> Result<ElementMeasures<F>, String> {
-    mesh.validate_nodes()?;
+    validate_axisymmetric_nodes(mesh)?;
     mesh.validate_connectivity()?;
     let mut areas = Vec::with_capacity(mesh.num_elements());
     let mut swept_volumes = Vec::with_capacity(mesh.num_elements());
@@ -283,7 +159,7 @@ fn element_quadrature_generic<F: Real, const NODES_PER_ELEMENT: usize>(
         QuadratureRule,
     ) -> Result<Vec<VolumeSample<F, NODES_PER_ELEMENT>>, String>,
 ) -> Result<ElementQuadrature<F>, String> {
-    mesh.validate_nodes()?;
+    validate_axisymmetric_nodes(mesh)?;
     mesh.validate_connectivity()?;
     let nq = quadrature.points_per_element();
     let mut points_rz = Vec::with_capacity(mesh.num_elements() * nq);
