+++ src/physics/solenoid_stress/types.rs
@@ -19,3 +19,54 @@ pub fn cast<F: Real>(value: f64) -> F {
 pub fn two_pi<F: Real>() -> F {
     cast(2.0 * core::f64::consts::PI)
 }
+
+/// Number of displacement unknowns carried by each node in the axisymmetric structural solver.
+pub const DOF_PER_NODE: usize = 2;
+
+/// Return the number of structural displacement unknowns carried by one element with the given
+/// node count.
+pub const fn dof_per_element(nodes_per_element: usize) -> usize {
+    DOF_PER_NODE * nodes_per_element
+}
+
+#[derive(Clone, Copy, Debug)]
+pub struct PressureLoad<F: Real> {
+    /// Element index receiving the load.
+    pub element: usize,
+    /// Local face index in the element-family numbering used by `face_reference`.
+    pub local_face: u8,
+    /// Pressure magnitude, taken positive in the inward normal direction.
+    pub value: F,
+}
+
+#[derive(Clone, Copy, Debug)]
+pub struct TractionLoad<F: Real> {
+    /// Element index receiving the load.
+    pub element: usize,
+    /// Local face index in the element-family numbering used by `face_reference`.
+    pub local_face: u8,
+    /// Constant traction vector in global meridian coordinates `[t_r, t_z]`.
+    pub value: [F; 2],
+}
+
+#[derive(Clone, Copy, Debug)]
+pub struct ThermalMaterial<F: Real> {
+    /// Thermal strain coefficients in axisymmetric strain order `[rr, zz, tt, rz]`.
+    pub alpha: [F; 4],
+    /// Stress-free reference temperature for this material.
+    pub reference_temperature: F,
+}
+
+#[derive(Debug, Clone)]
+pub struct AssemblyResult<F: Real> {
+    /// Sparse row indices for the assembled stiffness matrix triplets.
+    pub rows: Vec<usize>,
+    /// Sparse column indices for the assembled stiffness matrix triplets.
+    pub cols: Vec<usize>,
+    /// Sparse values for the assembled stiffness matrix triplets.
+    pub vals: Vec<F>,
+    /// Global right-hand side vector.
+    pub rhs: Vec<F>,
+    /// Total number of displacement unknowns in the global system.
+    pub ndof: usize,
+}
